requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic>=2.10.0
python-telegram-bot==21.6
gspread==6.1.2
google-auth>=2.29.0
instagrapi>=2.0.0
